backend/requirements.txt
azure-functions
openai
python-dotenv
